+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-#Version 0.1.0+5, (2014/12/02)
+#Version 0.1.0+6, (2014/12/24)
+- delay shadow dom component render till element style is ready
 
+#Version 0.1.0+5, (2014/12/10)
+- add column css sample
 ##Bug Fixes
 - fix numeric parsing